func-idgov-master/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
# pygsheets==2.0.6
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-monitor-query==1.2.0
msal==1.25.0
pandas==2.1.3
pyodbc==5.0.1
sqlalchemy==2.0.23
oauthlib==3.2.2
requests_oauthlib==1.3.1
